add components for managing external and internal marks, including view, entry, and edit functionalities
