+++ client/src/components/Teacher/ExternalMarksEntry.js
@@ -1,22 +1,20 @@
 import React, { useState, useEffect } from "react";
 import axios from "axios";
 
-function InternalMarks() {
+function ExternalMarksEntry() {
     const [marks, setMarks] = useState([]);
     const [formData, setFormData] = useState({
         studentName: "",
         subName: "",
         marks: ""
     });
-    const [editingMark, setEditingMark] = useState(null);
-    const [updatedMark, setUpdatedMark] = useState("");
 
     useEffect(() => {
         fetchMarks();
     }, []);
 
     const fetchMarks = async () => {
-        const response = await axios.get("http://localhost:5000/teacher/InternalMark");
+        const response = await axios.get("http://localhost:5000/teacher/ExternalMark");
         setMarks(response.data.data);
     };
 
@@ -26,32 +24,19 @@ function InternalMarks() {
 
     const handleSubmit = async (e) => {
         e.preventDefault();
-        await axios.post("http://localhost:5000/teacher/InternalMarkEntry", formData);
+        await axios.post("http://localhost:5000/teacher/ExternalMarkEntry", formData);
         setFormData({ studentName: "", subName: "", marks: "" });
         fetchMarks();
     };
 
-    const handleEdit = (mark) => {
-        setEditingMark(mark);
-        setUpdatedMark(mark.marks);
-    };
-
-    const handleUpdate = async () => {
-        if (!editingMark) return;
-        await axios.put(`http://localhost:5000/teacher/InternalMarkById/${editingMark._id}`, { marks: updatedMark });
-        setEditingMark(null);
-        setUpdatedMark("");
-        fetchMarks();
-    };
-
     const handleDelete = async (id) => {
-        await axios.delete(`http://localhost:5000/teacher/InternalMark/${id}`);
+        await axios.delete(`http://localhost:5000/teacher/ExternalMark/${id}`);
         fetchMarks();
     };
 
     return (
         <div className="p-4">
-            <h2 className="text-xl font-bold">Internal Marks</h2>
+            <h2 className="text-xl font-bold">External Marks</h2>
             <form onSubmit={handleSubmit} className="my-4">
                 <input type="text" name="studentName" placeholder="Student Name" value={formData.studentName} onChange={handleChange} className="border p-2 m-2" required />
                 <input type="text" name="subName" placeholder="Subject Name" value={formData.subName} onChange={handleChange} className="border p-2 m-2" required />
@@ -72,25 +57,9 @@ function InternalMarks() {
                         <tr key={mark._id}>
                             <td className="border p-2">{mark.studentName}</td>
                             <td className="border p-2">{mark.subName}</td>
+                            <td className="border p-2">{mark.marks}</td>
                             <td className="border p-2">
-                                {editingMark && editingMark._id === mark._id ? (
-                                    <input
-                                        type="number"
-                                        value={updatedMark}
-                                        onChange={(e) => setUpdatedMark(e.target.value)}
-                                        className="border p-1"
-                                    />
-                                ) : (
-                                    mark.marks
-                                )}
-                            </td>
-                            <td className="border p-2">
-                                {editingMark && editingMark._id === mark._id ? (
-                                    <button onClick={handleUpdate} className="bg-green-500 text-white p-1">Save</button>
-                                ) : (
-                                    <button onClick={() => handleEdit(mark)} className="bg-yellow-500 text-white p-1">Edit</button>
-                                )}
-                                <button onClick={() => handleDelete(mark._id)} className="bg-red-500 text-white p-1 ml-2">Delete</button>
+                                <button onClick={() => handleDelete(mark._id)} className="bg-red-500 text-white p-1">Delete</button>
                             </td>
                         </tr>
                     ))}
@@ -100,4 +69,4 @@ function InternalMarks() {
     );
 }
 
-export default InternalMarks;
+export default ExternalMarksEntry;
