+++ client/src/components/Teacher/ViewStatus.js
@@ -1,0 +1,9 @@
+import React from 'react'
+
+function viewStatus() {
+  return (
+    <div><h1>viewStatus</h1></div>
+  )
+}
+
+export default viewStatus
